+++ gems/kettle-jem/CHANGELOG.md
@@ -22,11 +22,21 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Changed
 
+- [kc] kettle-jem/template: updated 398 project files:
+  - code and tests (20)
+  - configuration (12)
+  - dependencies (147)
+  - documentation (9)
+  - other (123)
+  - workflows (87)
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- Preserve inherited version namespaces by declaring their superclass before loading version.rb.
 
 ### Security